pyproject.toml
[tool.poetry]
name = "new"
version = "0.1.0"
description = ""
authors = ["sergeytashkinov <tashkinov2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["uvicorn"], version = "^0.115.2"}
pydantic-settings = "^2.5.2"
sqlalchemy = "^2.0.35"
aiosqlite = "^0.20.0"
greenlet = "^3.1.1"
httpx = "^0.27.2"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
python-jose = {extras = ["cryptogrphy"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.2.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
python-multipart = "^0.0.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
